refactor: Moove IEnviromentVariables into common dir
